戦闘終了後アクション決定時の処理を別ファイルに切り出した
